add delete behavior for containers with objects

+++ swift_browser_ui_frontend/src/components/ContainerDeleteButton.vue
@@ -6,7 +6,6 @@
       outlined
       size="is-small"
       :inverted="inverted"
-      :disabled="disabled"
       @click="confirmDelete ()"
     >
       {{ $t('message.delete') }}
@@ -21,19 +20,33 @@ export default {
   name: "DeleteContainerButton",
   props: [
     "container",
-    "disabled",
     "inverted",
+    "objects",
   ],
   methods: {
     confirmDelete: function () {
-      this.$buefy.dialog.confirm({
-        title: this.$t("message.container_ops.deleteConfirm"),
-        message: this.$t("message.container_ops.deleteConfirmMessage"),
-        confirmText: this.$t("message.container_ops.deleteConfirm"),
-        type: "is-danger",
-        hasIcon: true,
-        onConfirm: () => {this.deleteContainer();},
-      });
+      if (this.$props.objects > 0) {
+        this.$buefy.notification.open({
+          message: "Deleting a container requires deleting all objects first.",
+          type: "is-danger",
+          position: "is-top-right",
+          hasIcon: true,
+        });
+        this.$router.push(
+          this.$route.params.project
+            + "/"
+            + this.$props.container,
+        );
+      } else {
+        this.$buefy.dialog.confirm({
+          title: this.$t("message.container_ops.deleteConfirm"),
+          message: this.$t("message.container_ops.deleteConfirmMessage"),
+          confirmText: this.$t("message.container_ops.deleteConfirm"),
+          type: "is-danger",
+          hasIcon: true,
+          onConfirm: () => {this.deleteContainer();},
+        });
+      }
     },
     deleteContainer: function () {
       this.$buefy.toast.open({
